copy files from sir to compiler-plugin only if one updated.

+++ build.sbt
@@ -131,8 +131,15 @@ lazy val scalusPlugin = project
               val target = targetDir / file
 
               if (source.exists) {
-                  IO.copyFile(source, target)
-                  log.info(s"Copied $file to target $target")
+                  if (!target.exists) {
+                      IO.copyFile(source, target)
+                      log.info(s"Copied $file to target $target")
+                  } else if (source.lastModified() > target.lastModified()) {
+                      IO.copyFile(source, target)
+                      log.info(s"Copied $file to target $target")
+                  } else {
+                      log.info(s"File $target is up to date")
+                  }
               } else {
                   log.error(s"Source file not found: $file")
               }
